Split utils into separate files; update imports
