+++ src/main/util.ts
@@ -1,4 +1,5 @@
-/* eslint import/prefer-default-export: off, import/no-mutable-exports: off */
+/* eslint import/no-mutable-exports: off */
+
 import { URL } from 'url';
 import path from 'path';
 import { existsSync, mkdirSync, statSync } from 'fs';
@@ -55,9 +56,3 @@ export const getThumbnailPath: (projectId: string) => string = (projectId) =>
 
 export const fileOrDirExists: (filePath: string) => boolean = (filePath) =>
   statSync(filePath, { throwIfNoEntry: false }) !== undefined;
-
-/** Utility types */
-
-// Callback to be passed into a map function.
-// First type argument is the input type, second is the output type
-export type MapCallback<T, U> = (val: T, index: number, arr: T[]) => U;
