+++ src/renderer/components/Word.tsx
@@ -7,7 +7,7 @@ import {
   RefObject,
 } from 'react';
 import { MousePosition } from '@react-hook/mouse-position';
-import { pointIsInsideRect } from 'renderer/util';
+import { pointIsInsideRect } from 'renderer/utils/geometry';
 import colors from '../colors';
 import { handleSelectWord } from '../editor/selection';
 import { DragState } from './WordDragManager';
