+++ src/renderer/utils/input.ts
@@ -1,0 +1,7 @@
+/* eslint-disable import/prefer-default-export */
+
+export enum MouseButton {
+  LEFT = 0,
+  MIDDLE = 1,
+  RIGHT = 2,
+}
